REFACTOR: refactored n_cusps calculation into _Cycloid ABC
